backend/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
pydantic>=2.0
python-multipart>=0.0.9
requests>=2.31
sse-starlette>=2.0

langchain>=0.2
langchain-huggingface>=0.1.0
sentence-transformers>=2.6.0
torch>=2.2.0
transformers>=4.40.0
langchain-community>=0.2
langchain-text-splitters>=0.2
pypdf>=4.0

faiss-cpu>=1.7.4
google-genai>=0.5
google-generativeai>=0.5
